rename to reflect stretch only
