+++ manipulathor_baselines/stretch_object_nav_baselines/experiments/ithor/obj_nav_2camera_ithor_wide.py
@@ -16,14 +16,14 @@ from allenact_plugins.ithor_plugin.ithor_sensors import GoalObjectTypeThorSensor
 
 from utils.stretch_utils.stretch_ithor_arm_environment import StretchManipulaTHOREnvironment
 
-from manipulathor_baselines.object_nav_baselines.experiments.obj_nav_base_config import ObjectNavBaseConfig
-from utils.procthor_utils.all_rooms_object_nav_task_sampler import AllRoomsObjectNavTaskSampler
-from utils.procthor_utils.procthor_object_nav_tasks import StretchObjectNavTask
+from manipulathor_baselines.stretch_object_nav_baselines.experiments.obj_nav_base_config import ObjectNavBaseConfig
+from utils.stretch_utils.all_rooms_object_nav_task_sampler import AllRoomsObjectNavTaskSampler
+from utils.stretch_utils.stretch_object_nav_tasks import StretchObjectNavTask
 from utils.stretch_utils.stretch_constants import STRETCH_MANIPULATHOR_COMMIT_ID, STRETCH_ENV_ARGS
 from manipulathor_utils.debugger_util import ForkedPdb
 
 # from manipulathor_baselines.procthor_baselines.models.clip_preprocessors import ClipResNetPreprocessor
-from manipulathor_baselines.object_nav_baselines.models.clip_resnet_ncamera_preprocess_mixin import \
+from manipulathor_baselines.stretch_object_nav_baselines.models.clip_resnet_ncamera_preprocess_mixin import \
     ClipResNetPreprocessNCameraGRUActorCriticMixin
 from allenact_plugins.clip_plugin.clip_preprocessors import ClipResNetPreprocessor
 from allenact.base_abstractions.preprocessor import Preprocessor
