+++ manipulathor_baselines/stretch_object_nav_baselines/models/clip_objnav_ncamera_model.py
